+++ verta/tests/test_datasets.py
@@ -16,6 +16,35 @@ from verta._protos.public.modeldb import DatasetService_pb2 as _DatasetService
 from verta._protos.public.modeldb import DatasetVersionService_pb2 as _DatasetVersionService
 
 
+DEFAULT_S3_TEST_BUCKET = "bucket"
+DEFAULT_S3_TEST_OBJECT = "object"
+DEFAULT_GOOGLE_APPLICATION_CREDENTIALS = "credentials.json"
+
+
+@pytest.fixture(scope='session')
+def s3_bucket():
+    return os.environ.get("VERTA_S3_TEST_BUCKET", DEFAULT_S3_TEST_BUCKET)
+
+
+@pytest.fixture(scope='session')
+def s3_object():
+    return os.environ.get("VERTA_S3_TEST_OBJECT", DEFAULT_S3_TEST_OBJECT)
+
+
+@pytest.fixture(scope='session')
+def bq_query():
+    return (
+        "SELECT id, `by`, score, time, time_ts, title, url, text, deleted, dead, descendants, author"
+        " FROM `bigquery-public-data.hacker_news.stories`"
+        " LIMIT 1000"
+    )
+
+
+@pytest.fixture(scope='session')
+def bq_location():
+    return "US"
+
+
 class TestBaseDatasets:
     def test_creation_from_scratch(self, client, created_datasets):
         name = utils.gen_str()
